remove unnecessary cast for file byte len

+++ src/counts.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 
 pub struct Counts {
-    pub bytes: usize,
+    pub bytes: u64,
     pub chars: usize,
     pub lines: usize,
     pub words: usize,
@@ -10,7 +10,7 @@ pub struct Counts {
 }
 
 impl Counts {
-    pub(crate) fn new(input_bytes: usize, input_chars: usize, input_lines: usize, input_words: usize, input_max_line: usize, input_file_path: PathBuf) -> Counts {
+    pub(crate) fn new(input_bytes: u64, input_chars: usize, input_lines: usize, input_words: usize, input_max_line: usize, input_file_path: PathBuf) -> Counts {
         Counts { bytes: input_bytes, chars: input_chars, lines: input_lines, words: input_words, max_line: input_max_line, file_path: input_file_path }
     }
 }
